+++ packages/react-core/src/hooks/index.ts
@@ -1,2 +1,6 @@
-export { default as usePreviousValue } from './usePreviousValue';
+export {
+  default as useDeprecationWarning,
+  UseDeprecationWarning,
+} from './useDeprecationWarning';
 export { default as useHasValueUpdated } from './useHasValueUpdated';
+export { default as usePreviousValue } from './usePreviousValue';