+++ packages/react-liveness/src/components/FaceLivenessDetector/service/machine/__tests__/index.test.ts
@@ -163,7 +163,6 @@ describe('Liveness Machine', () => {
     jest.advanceTimersToNextTimer(); // waitForSessionInformation
     await flushPromises(); // detectFaceDistanceBeforeRecording
     jest.advanceTimersToNextTimer(); // checkFaceDistanceBeforeRecording
-    service.send({ type: 'START_RECORDING' });
   }
 
   async function advanceMinFaceMatches() {
@@ -193,6 +192,7 @@ describe('Liveness Machine', () => {
 
     mockedHelpers.isCameraDeviceVirtual.mockImplementation(() => false);
     mockedHelpers.VideoRecorder.mockImplementation(() => mockVideoRecorder);
+    (mockVideoRecorder.getVideoChunkSize as jest.Mock).mockReturnValue(10);
     mockedHelpers.BlazeFaceFaceDetection.mockImplementation(
       () => mockBlazeFace
     );
@@ -420,7 +420,7 @@ describe('Liveness Machine', () => {
   describe('notRecording', () => {
     it('should reach recording state on START_RECORDING', async () => {
       await transitionToInitializeLivenessStream(service);
-      await flushPromises(); // notRecording
+      await flushPromises(); // checkFaceDistanceBeforeRecording
 
       service.send({
         type: 'SET_SESSION_INFO',
@@ -428,10 +428,9 @@ describe('Liveness Machine', () => {
           sessionInfo: mockSessionInformation,
         },
       });
-      jest.advanceTimersToNextTimer(); // waitForSessionInformation
-      await flushPromises(); // detectFaceDistanceBeforeRecording
-      jest.advanceTimersToNextTimer(); // checkFaceDistanceBeforeRecording
-      service.send({ type: 'START_RECORDING' });
+      jest.advanceTimersToNextTimer(); // initializeLivenessStream
+      await flushPromises(); // { notRecording: 'waitForSessionInfo' }
+      jest.advanceTimersToNextTimer(); // { recording: 'ovalDrawing' }
 
       expect(service.state.value).toEqual({ recording: 'ovalDrawing' });
     });
@@ -496,7 +495,6 @@ describe('Liveness Machine', () => {
         .mockResolvedValue([mockFace])
         .mockResolvedValueOnce([mockFace]) // first to pass detecting face before start
         .mockResolvedValueOnce([mockFace]) // second to pass face distance before start
-        .mockResolvedValueOnce([mockFace]) // third to pass face distance check after countdown
         .mockResolvedValueOnce([]); // not having face in view when recording begins
       mockedHelpers.estimateIllumination.mockImplementation(
         () => IlluminationState.BRIGHT
@@ -523,7 +521,6 @@ describe('Liveness Machine', () => {
         .mockResolvedValue([mockFace])
         .mockResolvedValueOnce([mockFace]) // first to pass detecting face before start
         .mockResolvedValueOnce([mockFace]) // second to pass face distance before start
-        .mockResolvedValueOnce([mockFace]) // third to pass face distance check after countdown
         .mockRejectedValue(error);
 
       await transitionToRecording(service);
@@ -728,6 +725,23 @@ describe('Liveness Machine', () => {
       const livenessError = (mockcomponentProps.onError as jest.Mock).mock
         .calls[0][0];
       expect(livenessError.state).toBe(LivenessErrorState.SERVER_ERROR);
+    });
+
+    it('should reach error state if no chunks are recorded', async () => {
+      const error = new Error('Video chunks not recorded successfully.');
+      error.name = LivenessErrorState.RUNTIME_ERROR;
+
+      (mockVideoRecorder.getVideoChunkSize as jest.Mock).mockReturnValue(0);
+      await transitionToUploading(service);
+
+      await flushPromises(); // stopVideo
+
+      expect(service.state.value).toEqual('error');
+      expect(service.state.context.errorState).toBe(
+        LivenessErrorState.RUNTIME_ERROR
+      );
+      expect(mockcomponentProps.onError).toHaveBeenCalledTimes(1);
+      expect(mockcomponentProps.onError).toHaveBeenCalledWith(error);
     });
   });
 });