+++ packages/react-liveness/src/components/FaceLivenessDetector/service/machine/index.ts
@@ -121,6 +121,9 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
         target: 'error',
         actions: 'updateErrorStateForServer',
       },
+      RUNTIME_ERROR: {
+        target: 'error',
+      },
       MOBILE_LANDSCAPE_WARNING: {
         target: 'mobileLandscapeWarning',
         actions: 'updateErrorStateForServer',
@@ -199,7 +202,7 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
         invoke: {
           src: 'openLivenessStreamConnection',
           onDone: {
-            target: 'waitForSessionInfo',
+            target: 'notRecording',
             actions: [
               'updateLivenessStreamProvider',
               'spawnResponseStreamActor',
@@ -207,40 +210,40 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
           },
         },
       },
-      waitForSessionInfo: {
-        after: {
-          0: {
-            target: 'notRecording',
-            cond: 'hasServerSessionInfo',
-          },
-          100: { target: 'waitForSessionInfo' },
-        },
-      },
       notRecording: {
         on: {
           START_RECORDING: 'recording', // if countdown completes while face is far enough, start recording
         },
-        initial: 'detectFaceDistanceDuringCountdown',
+        initial: 'waitForSessionInfo',
         states: {
-          detectFaceDistanceDuringCountdown: {
+          waitForSessionInfo: {
+            after: {
+              0: {
+                target: '#livenessMachine.recording',
+                cond: 'hasServerSessionInfo',
+              },
+              100: { target: 'detectFaceDistanceDuringLoading' },
+            },
+          },
+          detectFaceDistanceDuringLoading: {
             invoke: {
-              src: 'detectFaceDistance',
+              src: 'detectFaceDistanceWhileLoading',
               onDone: {
-                target: 'checkFaceDistanceDuringCountdown',
-                actions: ['updateFaceDistanceBeforeRecording'],
+                target: 'checkFaceDistanceDuringLoading',
+                actions: ['updateFaceDistanceWhileLoading'],
               },
             },
           },
-          checkFaceDistanceDuringCountdown: {
-            after: {
-              0: {
+          checkFaceDistanceDuringLoading: {
+            always: [
+              {
                 target: 'failure',
                 cond: 'hasNotEnoughFaceDistanceBeforeRecording',
               },
-              200: {
-                target: 'detectFaceDistanceDuringCountdown',
+              {
+                target: 'waitForSessionInfo',
               },
-            },
+            ],
           },
           failure: {
             entry: 'sendTimeoutAfterFaceDistanceDelay',
@@ -566,6 +569,14 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
           return event.data!.isFaceFarEnoughBeforeRecording;
         },
       }),
+      updateFaceDistanceWhileLoading: assign({
+        isFaceFarEnoughBeforeRecording: (_, event) => {
+          return event.data!.isFaceFarEnoughBeforeRecording;
+        },
+        errorState: (_, event) => {
+          return event.data?.error;
+        },
+      }),
       updateOvalAndFaceDetailsPostDraw: assign({
         ovalAssociatedParams: (context, event) => ({
           ...context.ovalAssociatedParams,
@@ -619,7 +630,9 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
         },
       }),
       updateErrorStateForRuntime: assign({
-        errorState: (_) => LivenessErrorState.RUNTIME_ERROR,
+        errorState: (_, event) => {
+          return event.data?.errorState || LivenessErrorState.RUNTIME_ERROR;
+        },
       }),
       updateErrorStateForServer: assign({
         errorState: (_) => LivenessErrorState.SERVER_ERROR,
@@ -695,8 +708,7 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
       ),
       sendTimeoutAfterFaceDistanceDelay: actions.send(
         {
-          type: 'TIMEOUT',
-          data: { errorState: LivenessErrorState.FACE_DISTANCE_ERROR },
+          type: 'RUNTIME_ERROR',
         },
         {
           delay: 0,
@@ -950,7 +962,7 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
           height: height!,
         });
 
-        const isFaceFarEnoughBeforeRecording =
+        const { isDistanceBelowThreshold: isFaceFarEnoughBeforeRecording } =
           await isFaceDistanceBelowThreshold({
             faceDetector: faceDetector!,
             videoEl: videoEl!,
@@ -960,6 +972,36 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
           });
 
         return { isFaceFarEnoughBeforeRecording };
+      },
+      async detectFaceDistanceWhileLoading(context) {
+        const {
+          isFaceFarEnoughBeforeRecording: faceDistanceCheckBeforeRecording,
+        } = context;
+        const { videoEl, videoMediaStream, isMobile } =
+          context.videoAssociatedParams!;
+        const { faceDetector } = context.ovalAssociatedParams!;
+
+        const { width, height } = videoMediaStream!
+          .getTracks()[0]
+          .getSettings();
+
+        const ovalDetails = getStaticLivenessOvalDetails({
+          width: width!,
+          height: height!,
+        });
+
+        const {
+          isDistanceBelowThreshold: isFaceFarEnoughBeforeRecording,
+          error,
+        } = await isFaceDistanceBelowThreshold({
+          faceDetector: faceDetector!,
+          videoEl: videoEl!,
+          ovalDetails,
+          reduceThreshold: faceDistanceCheckBeforeRecording, // if this is the second face distance check reduce the threshold
+          isMobile,
+        });
+
+        return { isFaceFarEnoughBeforeRecording, error };
       },
       async detectInitialFaceAndDrawOval(context) {
         const { serverSessionInformation, livenessStreamProvider } = context;
@@ -1178,15 +1220,16 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
           },
         };
 
+        if (livenessStreamProvider!.videoRecorder.getVideoChunkSize() === 0) {
+          throw new Error('Video chunks not recorded successfully.');
+        }
+
         livenessStreamProvider!.sendClientInfo(livenessActionDocument);
 
         await livenessStreamProvider!.dispatchStopVideoEvent();
       },
       async getLiveness(context) {
-        const { livenessStreamProvider } = context;
         const { onAnalysisComplete } = context.componentProps!;
-
-        livenessStreamProvider!.endStream();
 
         // Get liveness result
         await onAnalysisComplete();