+++ packages/react-core-notifications/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @aws-amplify/ui-react-core-notifications
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`b0e16e78c`](https://github.com/aws-amplify/amplify-ui/commit/b0e16e78c6a41945aa79f3e14fa3f9e6cb0e5e76), [`d930e2ed1`](https://github.com/aws-amplify/amplify-ui/commit/d930e2ed17f3e638e2b62699ba2dd164b32f8118)]:
+  - @aws-amplify/ui@5.6.6
+  - @aws-amplify/ui-react-core@2.1.25
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`23180b470`](https://github.com/aws-amplify/amplify-ui/commit/23180b470c7b3b78a5970d00f8c2dc5ce8773eff), [`3cecd0765`](https://github.com/aws-amplify/amplify-ui/commit/3cecd0765b46c77c49af24fae7cfb9054ebe2cdb)]:
+  - @aws-amplify/ui-react-core@2.1.24
+  - @aws-amplify/ui@5.6.5
+
 ## 1.0.0
 
 ### Major Changes